Explicitly point PUSH instructions with OFFSETs into BSS to DGROUP
What the hell.

So, with the COMPACT model, we once again run into the issue of c0's
_TEXT segment ending on an odd byte (0x263) and our code immediately
coming after. This means we can't use .MODEL and the simplified segment
directives (.CODE, .DATA, and .DATA?), because TASM enforces word
alignment between c0's _TEXT and our _TEXT in every single memory model,
and redefining that alignment is somehow illegal and an error and TASM
absolutely can't compile code like that. -.-

Well, it actually isn't much of a problem, because we can simply declare
segments with their proper names and our intended alignment ourselves,
put the correct segments into DGROUP, and everything works out just fine.

… *except* for PUSH OFFSET instructions if the offset points into BSS,
where the offset stays relative to BSS in the final binary, despite BSS
being in DGROUP and DS being ASSUMEd as DGROUP. This *has* to be a bug
in TASM or TLINK,  every other reference to those BSS pointers is
calculated relative to DGROUP just fine.

And since we can't emulate .MODEL by manually defining all symbols it
defines, we just have to manually point every one of those offsets to
DGROUP. Oh well.