[Reduction] Identify and remove the C runtime
It's amazing how almost impossible it was to find *any* version of
MAKER.EXE whose libc exactly matches *any* version of a Borland
compiler. So, Borland C++ 5.01 it is, exactly. 4.52 is too old, and 5.02
is too new.

And yay! We can indeed debug MAKER40D.EXE in Turbo Debugger now.